Intermediate commit to clean up warning messages

+++ src/ofxBox2dPolygon.cpp
@@ -49,7 +49,7 @@ void ofxBox2dPolygon::addTriangle(const ofVec2f &a, const ofVec2f &b, const ofVe
 */
 //----------------------------------------
 void ofxBox2dPolygon::addVertexes(vector <ofVec2f> &pts) {
-	for (int i=0; i<pts.size(); i++) {
+	for (size_t i=0; i<pts.size(); i++) {
 		addVertex(pts[i].x, pts[i].y);
 	}
 }
